Logs ingest newpath (#87)
This is full refactor of the data processing pipeline to have composable readers, memory management, and correctness checks along the path to ensure we do not corrupt data, the input is consistent, and specific characteristics of the data can be manipulated along the path.

For log and metric ingest, we now support `binpb` and `binpb.gz` format.

Metrics are written to our "cooked" Parquet format in a sorted order now, so further work on these should not require any expensive sorting operations for aggregation and compaction.  A merge sort is sufficient.

Traces are not part of this path currently, nor are log compaction.  These will be upgraded in the future.